Rewrite the build system.
This heavily relies on buildSrc and conventions to unify most of the build architecture. Versions are now externally defined via TOML file. Style is enforced via Spotless, but styling has not been applied to any files not touched by this commit. Documentation is now a gradle project with custom Jekyll build and serve tasks.