gui/util/BidirectionalMap: Cleanup, fix, tests
One notable change is the adjustment of functionality in addAll. Before, it was possible to create a map with duplicate keys/values by adding them twice using the same `addAll` call, as there were no check within that function call whether they are even valid in that sense. Apart from that, a few tests have been added to get the coverage to 100%.